[Feature] CLDR Plural Rules Support with Decimal Number Pluralization (#43)
* feat: Update TranslationController and related components to support decimal numbers for pluralization

- Modified the `number` parameter in `GetEntry` and `GetEntryUsingClient` methods to accept decimal values, enhancing pluralization support for fractional numbers.
- Updated XML documentation to clarify that the `number` parameter now supports both integer and decimal values.
- Adjusted related interfaces and classes across the SDK to ensure consistent handling of decimal numbers for translation entries.
- Added unit tests to validate the new functionality and ensure correct behavior with decimal inputs.

* fix: Update test setups to use decimal instead of int for translation methods

- Modified unit tests in CachingTranslaasClientTests, TranslaasHelperTests, and TranslaasTagHelperTests to replace int? with decimal? for the number parameter in translation method setups.
- Ensured consistency with recent changes to support decimal values for pluralization in the Translaas SDK.

* feat: Enhance Translaas SDK configuration and add sample appsettings files

- Added Microsoft.Extensions.Configuration.UserSecrets package to support user secrets management.
- Introduced appsettings.json and appsettings.Development.json files for configuration settings, including Translaas API base URL, caching options, and logging levels.
- Updated Program.cs to load user secrets and validate API key configuration, providing error messages for missing keys.
- Enhanced Directory.Packages.props to include the new UserSecrets package reference.

* refactor: Simplify caching configuration in Program.cs

- Streamlined the configuration of CacheAbsoluteExpiration, CacheSlidingExpiration, and Timeout settings by consolidating conditional checks into a single line using the ternary operator.
- Improved code readability and maintainability by reducing repetitive code patterns in the caching setup.

* refactor: Optimize plural forms parsing in TranslationGroup

- Streamlined the parsing of plural forms in the TranslationGroup class by replacing the manual iteration with LINQ expressions.
- Improved code readability and maintainability by reducing the number of lines and enhancing the logic for handling plural categories and values.

+++ src/Translaas.Caching.File/CachingTranslaasClient.cs
@@ -44,7 +44,7 @@ public class CachingTranslaasClient : ITranslaasClient
         string group,
         string entry,
         string lang,
-        int? number = null,
+        decimal? number = null,
         CancellationToken cancellationToken = default)
     {
         return _options.FallbackMode switch
@@ -113,7 +113,7 @@ public class CachingTranslaasClient : ITranslaasClient
         string group,
         string entry,
         string lang,
-        int? number,
+        decimal? number,
         CancellationToken cancellationToken)
     {
         // Try cache first
@@ -145,7 +145,7 @@ public class CachingTranslaasClient : ITranslaasClient
         string group,
         string entry,
         string lang,
-        int? number,
+        decimal? number,
         CancellationToken cancellationToken)
     {
         try
@@ -193,7 +193,7 @@ public class CachingTranslaasClient : ITranslaasClient
         string group,
         string entry,
         string lang,
-        int? number,
+        decimal? number,
         CancellationToken cancellationToken)
     {
         var result = await _innerClient.GetEntryAsync(group, entry, lang, number, cancellationToken).ConfigureAwait(false);
